Use copy site|lang syntax
